docs: split sidebars by route-group and add icon mapping; blog: add TS/@types/eslint devDeps for Vercel build
