+++ apps/blog/package.json
@@ -25,5 +25,13 @@
     "remark-gfm": "^4.0.1",
     "remark-mdx": "^3.1.0",
     "tailwind-merge": "^3.3.1"
+  },
+  "devDependencies": {
+    "@types/node": "24.0.13",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "eslint": "^8",
+    "eslint-config-next": "15.4.1",
+    "typescript": "^5.8.3"
   }
 }
