+++ apps/docs/app/(community)/layout.tsx
@@ -3,10 +3,52 @@ import type { ReactNode } from 'react';
 import { source } from '@/lib/source';
 import * as Lucide from 'lucide-react';
 
+type TreeNode = {
+  type?: string;
+  name?: string;
+  url?: string;
+  icon?: unknown;
+  children?: TreeNode[];
+};
+
+function toPascalCase(input: string): string {
+  return input
+    .split(/[\s-_]+/)
+    .map((s) => s.charAt(0).toUpperCase() + s.slice(1))
+    .join('');
+}
+
+const iconAlias: Record<string, string> = {
+  aid: 'Globe',
+  docs: 'Book',
+  documentation: 'Book',
+  book: 'Book',
+  spec: 'FileText',
+  specification: 'FileText',
+  rationale: 'HelpCircle',
+  versioning: 'GitBranch',
+  changelog: 'GitBranch',
+  quickstart: 'Rocket',
+  guide: 'FileText',
+};
+
+function mapIcons(node: TreeNode): TreeNode {
+  const mapped: TreeNode = { ...node };
+  if (typeof node.icon === 'string') {
+    const aliased = iconAlias[node.icon.toLowerCase()] ?? node.icon;
+    const pascal = toPascalCase(aliased);
+    type IconComponent = React.ComponentType<{ className?: string }>;
+    const IconComp = (Lucide as unknown as Record<string, IconComponent>)[pascal];
+    if (IconComp) mapped.icon = <IconComp className="size-4" />;
+  }
+  if (Array.isArray(node.children)) mapped.children = node.children.map(mapIcons);
+  return mapped;
+}
+
 export default function CommunityLayout({ children }: { children: ReactNode }) {
   return (
     <DocsLayout
-      tree={source.pageTree as any}
+      tree={mapIcons(source.pageTree as unknown as TreeNode) as Parameters<typeof DocsLayout>[0]['tree']}
       nav={{ enabled: true }}
       sidebar={{
         defaultOpenLevel: 0,
